COMMERCE-7567 batch-planner-web - reuse search bar configuration from Export/Import tab + use new search methods
