+++ modules/apps/batch-planner/batch-planner-web/src/main/java/com/liferay/batch/planner/web/internal/display/context/BaseSearchContainerManagementToolbarDisplayContext.java
@@ -14,15 +14,24 @@
 
 package com.liferay.batch.planner.web.internal.display.context;
 
+import com.liferay.batch.planner.model.BatchPlannerPlanTable;
 import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.SearchContainerManagementToolbarDisplayContext;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenuBuilder;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemBuilder;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
 import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
+
+import java.util.List;
+import java.util.Objects;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -76,6 +85,60 @@ public abstract class BaseSearchContainerManagementToolbarDisplayContext
 					LanguageUtil.get(httpServletRequest, "import-file"));
 			}
 		).build();
+	}
+
+	@Override
+	public List<DropdownItem> getFilterDropdownItems() {
+		List<DropdownItem> searchByDropdownItems = _getSearchByDropdownItems();
+
+		DropdownItemList filterDropdownItems = DropdownItemListBuilder.addAll(
+			super.getFilterDropdownItems()
+		).addGroup(
+			() -> searchByDropdownItems != null,
+			dropdownGroupItem -> {
+				dropdownGroupItem.setDropdownItems(searchByDropdownItems);
+				dropdownGroupItem.setLabel(_getSearchByDropdownItemsLabel());
+			}
+		).build();
+
+		if (filterDropdownItems.isEmpty()) {
+			return null;
+		}
+
+		return filterDropdownItems;
+	}
+
+	private DropdownItem _getSearchByDropdownItem(
+		String fieldName, String label) {
+
+		return DropdownItemBuilder.setActive(
+			Objects.equals(_getSearchByField(), fieldName)
+		).setHref(
+			getPortletURL(), "searchByField", fieldName
+		).setLabel(
+			LanguageUtil.get(httpServletRequest, label)
+		).build();
+	}
+
+	private List<DropdownItem> _getSearchByDropdownItems() {
+		return DropdownItemListBuilder.add(
+			_getSearchByDropdownItem(
+				BatchPlannerPlanTable.INSTANCE.name.getName(), "name")
+		).add(
+			_getSearchByDropdownItem(
+				BatchPlannerPlanTable.INSTANCE.internalClassName.getName(),
+				"entity")
+		).build();
+	}
+
+	private String _getSearchByDropdownItemsLabel() {
+		return LanguageUtil.get(httpServletRequest, "search-by");
+	}
+
+	private String _getSearchByField() {
+		return ParamUtil.getString(
+			liferayPortletRequest, "searchByField",
+			BatchPlannerPlanTable.INSTANCE.name.getName());
 	}
 
 }
